Adapted economic model to use a lockdown timeseries
